update readme;Rename AggAssistantListExtractor to AggListExtractor;Rename AggAssistantObjectExtractor to AggObjectExtractor; Add comment; Move construct method from AggAssistant to echo sub Assistant.
