+++ src/main/java/org/springframework/data/elasticsearch/core/aggregation/AggregationsResultsExtractor.java
@@ -4,6 +4,9 @@ import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.search.aggregations.Aggregations;
 import org.springframework.data.elasticsearch.core.ResultsExtractor;
 
+/**
+ * This Extractor will check error and only return the aggregation in response.
+ */
 public class AggregationsResultsExtractor implements ResultsExtractor<Aggregations> {
     @Override
     public Aggregations extract(SearchResponse response) {
